Separate KV v1 and v2 into different files
